fix: lifetimes and generics (#36)
* fix: split lifetimes in queries and params

Previously it was not possible to return async streams created using
`.iter()` from functions which receive the client as an argument and
create the statement or parameters locally, because they would all
have to share the same lifetime `'a`. Using different lifetimes for the
client, statement and parameters should solve this problem.

* fix: remove type parameter `P` in `query_raw`

Previously it was not possible to use the `.iter()` and `.all()` methods
in functions which had to return an `impl Future<Output = ...> + Send`.
In this case the compiler would complain about a type mismatch caused
by a lifetime requirement introduced in the `Item = P` bound in the
`query_raw` declaration. Removing the type parameter `P` and moving its
bounds to `I::Item` seems to solve the problem.